New env var EXIT_STATUS to signal content failure
This implements a new environment variable `EXIT_STATUS`, which will
signal a failure if the _content_ of one of the feeds is malformed, but
still viable. This is implemented to prevent missing information in the
community lessons feed from causing errors building our entire feeds
structure.

If there are errors in community lessons, we will modify the
`EXIT_STATUS` environment variable to be `1` so that healthchecks pings
an error, even though the build itself will succeed. This means that we
are alerted when things need to be fixed, but we have some grace as to
how they can be fixed.