workflow_dispatch on to allow manual run of action
In addition to having scheduled actions, I would like us to have the ability to manually run an action (so we don't have to wait for the next scheduled run if we fix a bug or make a change in between).  According to the instructions on how to manually run a workflow,  it says workflow_dispatch must be on.  I would like to have someone who knows more about GH Actions than me to review this and be sure it won't cause a conflict with the scheduled runs.